Refactor Field and FieldValue.
