+++ src/codeschool/core/users/api.py
@@ -1,5 +1,5 @@
 from codeschool.api import router
 from . import views
 
-router.register(r'users', views.UserViewSet,base_name="users")
+router.register(r'users', views.UserViewSet)
 router.register(r'profile', views.ProfileViewSet)
