--- src/codeschool/lms/field/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class FieldConfig(AppConfig):
-    name = 'codeschool.lms.field'
