--- src/codeschool/lms/field_value/api.py
@@ -1,4 +1,0 @@
-from codeschool.api import router
-from . import views
-
-router.register('field_values', views.FieldValueViewSet)
