--- src/codeschool/lms/field_value/apps.py
@@ -1,4 +1,0 @@
-from django.apps import AppConfig
-
-class FieldValueConfig(AppConfig):
-    name = 'codeschool.lms.field_value'
