--- src/codeschool/lms/field_value/models.py
@@ -1,14 +1,0 @@
-from codeschool import models
-from codeschool.lms.field.models import Field
-from codeschool.core.users.models import User
-
-class FieldValue(models.Model):
-
-    content = models.CharField(
-        max_length = 200,
-    )
-
-    # Field created in class FieldModel
-    fields = models.ManyToManyField(Field)
-
-    user = models.ManyToManyField(User)
