--- src/codeschool/lms/field_value/templates/field_value.html
@@ -1,10 +1,0 @@
-{% block content %}
-
-<h2> Field Value </h2>
-
-<form method="post">
-  {% csrf_token %}
-  {{ form.as_p }}
-</form>
-
-{% endblock %}
