--- src/codeschool/lms/field_value/views.py
@@ -1,12 +1,0 @@
-from rest_framework import viewsets
-
-from . import serializers
-from .models import FieldValue
-
-class FieldValueViewSet(viewsets.ModelViewSet):
-    """
-    List of Field Values
-    """
-
-    queryset = FieldValue.objects.all()
-    serializer_class = serializers.FieldValueSerializer
