+++ src/codeschool/lms/fields/api.py
@@ -2,3 +2,4 @@ from codeschool.api import router
 from . import views
 
 router.register('fields', views.FieldViewSet)
+router.register('field_values', views.FieldValueViewSet)
