+++ src/codeschool/lms/fields/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class FieldsConfig(AppConfig):
+    name = 'codeschool.lms.fields'
