+++ src/codeschool/lms/fields/views.py
@@ -1,7 +1,7 @@
 from rest_framework import viewsets
 
 from . import serializers
-from .models import Field
+from .models import Field, FieldValue
 
 class FieldViewSet(viewsets.ModelViewSet):
     """
@@ -10,3 +10,11 @@ class FieldViewSet(viewsets.ModelViewSet):
 
     queryset = Field.objects.all()
     serializer_class = serializers.FieldSerializer
+
+class FieldValueViewSet(viewsets.ModelViewSet):
+    """
+    List of Field Values
+    """
+
+    queryset = FieldValue.objects.all()
+    serializer_class = serializers.FieldValueSerializer
